auth fixed

+++ src/controllers/blog.controller.ts
@@ -3,8 +3,21 @@ import { Blog } from "../models/blog.model.ts";
 
 export const getBlogs = async (req: Request, res: Response) => {
   try {
-    const blogs = await Blog.find();
-    return res.status(200).json(blogs);
+    const { search = "", page = 0, limit = 20 } = req.query;
+    const pageNumber = Number(page);
+    const limitNumber = Number(limit);
+    const query = {
+      $or: [
+        { title: { $regex: search as string, $options: "i" } },
+        { description: { $regex: search as string, $options: "i" } },
+      ],
+    };
+    const total = await Blog.countDocuments(query);
+    const totalPages = Math.ceil(total / limitNumber);
+    const blogs = await Blog.find(query)
+      .skip(pageNumber * limitNumber)
+      .limit(limitNumber);
+    return res.status(200).json({ blogs, total, totalPages });
   } catch (error) {
     res.status(500).json({ message: "Internal Server Error", error });
   }
@@ -13,6 +26,9 @@ export const getBlogs = async (req: Request, res: Response) => {
 export const getBlog = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ message: "Details required" });
+    }
     const blog = await Blog.findById(id);
     if (!blog) {
       return res.status(404).json({ message: "Blog not found" });
@@ -36,6 +52,9 @@ export const createBlog = async (req: Request, res: Response) => {
 export const updateBlog = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ message: "Details required" });
+    }
     const blog = await Blog.findById(id);
     if (!blog) {
       return res.status(404).json({ message: "Blog not found" });
@@ -53,9 +72,12 @@ export const updateBlog = async (req: Request, res: Response) => {
 export const deleteBlog = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    if (!id) {
+      return res.status(400).json({ message: "Details required" });
+    }
     const blog = await Blog.findById(id);
     if (!blog) {
-      return res.status(404).json({ message: "Blog not found" });
+      return res.status(404).json({ message: "Blog does not exist" });
     }
     await Blog.findByIdAndDelete(id);
     res.status(200).json({ message: "Blog deleted successfully" });
